Lots of GUI work

+++ GoTvInterface/IGoTv.cpp
@@ -178,7 +178,7 @@ void ff_avutil_log(void* ptr, int level, const char* format, va_list va)
 
     buffer += message;
     int pos, start = 0;
-    while( buffer.length() && ( pos = buffer.find_first_of( '\n', start ) ) >= 0 )
+    while( buffer.length() && ( pos = (int)buffer.find_first_of( '\n', start ) ) >= 0 )
     {
         if( pos > start )
         {
